requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8
-e git+https://github.com/rivayama/django-cronjobs.git@a0b7ed93cdab6c3c4906d9f8c0fc1e162adb0fb1#egg=django_cronjobs-master
django-dotenv==1.3.0
django-pgjson==0.3.0
django-toolbelt==0.0.1
gunicorn==19.3.0
oauthlib==0.7.2
psycopg2==2.6
requests==2.7.0
requests-oauthlib==0.5.0
static3==0.6.0
whitenoise==1.0.6
